pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlx-turboquant"
version = "1.0.0"
description = "Benchmarking and integration layer for TurboQuant KV cache compression on Apple Silicon"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Alexis Castellanos"},
]
keywords = ["mlx", "turboquant", "kv-cache", "benchmark", "apple-silicon", "tool-calling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mlx>=0.20.0",
    "mlx-lm>=0.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
turboquant = ["mlx-vlm>=0.1.0"]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
]
viz = ["matplotlib>=3.7.0"]

[project.urls]
Homepage = "https://github.com/alex-rentel/mlx-turboquant"
Repository = "https://github.com/alex-rentel/mlx-turboquant"

[tool.setuptools.packages.find]
include = ["benchmarks*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
